backend/requirements.txt
Django>=4.0,<4.1
djangorestframework>=3.14.0
psycopg2-binary>=2.9.5
django-filter>=22.1
django-cors-headers>=3.13.0
celery>=5.2.7
redis>=4.4.0
gunicorn>=20.1.0
python-dotenv>=0.21.0
scikit-learn>=1.1.3
xgboost>=1.7.1
shap>=0.41.0
geopandas>=0.12.1
rasterio>=1.3.3
xarray>=2022.11.0
netCDF4>=1.6.2
pandas>=1.5.1
numpy>=1.23.4
matplotlib>=3.6.2
joblib>=1.2.0